requirements.txt
# Core dependencies
arxiv>=1.4.0
requests>=2.28.0
PyYAML>=6.0
feedparser>=6.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.20.0

# OCR dependencies
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
pillow>=9.0.0

# For Nougat OCR
nougat-ocr>=0.1.0

# For Unstructured OCR
unstructured>=0.10.0
unstructured[pdf]>=0.10.0

# For Surya OCR
surya-ocr>=0.4.0

# LLM integration
openai>=1.0.0
anthropic>=0.3.0
mistralai>=1.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Text processing
nltk>=3.8.0
spacy>=3.4.0
regex>=2022.7.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development tools
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Logging
loguru>=0.6.0

# Configuration
pydantic>=1.10.0

# CLI and table formatting
tabulate>=0.9.0
questionary>=2.0.0
# For Research Assistant
pocketflow
litellm
questionary